pyproject.toml
[tool.poetry]
name = "learnotpyrc"
version = "0.1.0"
description = "tempProject"
authors = ["nohvenell2 <nohvenell2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyjwt = "^2.10.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
pytz = "^2025.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
